# sdk/typescript-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ChallengeCreateRequest": {
        "properties": {
          "complexity_slider": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Complexity Slider",
            "type": "number"
          },
          "iteration_window_seconds": {
            "exclusiveMinimum": 0.0,
            "maximum": 86400.0,
            "title": "Iteration Window Seconds",
            "type": "integer"
          },
          "minimum_quality_threshold": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Minimum Quality Threshold",
            "type": "number"
          },
          "prompt": {
            "minLength": 10,
            "title": "Prompt",
            "type": "string"
          },
          "risk_appetite": {
            "enum": [
              "conservative",
              "balanced",
              "aggressive"
            ],
            "title": "Risk Appetite",
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "prompt",
          "iteration_window_seconds",
          "minimum_quality_threshold",
          "risk_appetite",
          "complexity_slider"
        ],
        "title": "ChallengeCreateRequest",
        "type": "object"
      },
      "ChallengeResponse": {
        "properties": {
          "complexity_slider": {
            "title": "Complexity Slider",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "iteration_window_seconds": {
            "title": "Iteration Window Seconds",
            "type": "integer"
          },
          "minimum_quality_threshold": {
            "title": "Minimum Quality Threshold",
            "type": "number"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "risk_appetite": {
            "title": "Risk Appetite",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "prompt",
          "iteration_window_seconds",
          "minimum_quality_threshold",
          "risk_appetite",
          "complexity_slider",
          "created_at",
          "updated_at"
        ],
        "title": "ChallengeResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JudgeProfileInput": {
        "properties": {
          "domain": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Domain",
            "type": "string"
          },
          "head_judge": {
            "default": false,
            "title": "Head Judge",
            "type": "boolean"
          },
          "profile_prompt": {
            "minLength": 8,
            "title": "Profile Prompt",
            "type": "string"
          },
          "scoring_style": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Scoring Style",
            "type": "string"
          }
        },
        "required": [
          "domain",
          "scoring_style",
          "profile_prompt"
        ],
        "title": "JudgeProfileInput",
        "type": "object"
      },
      "JudgeProfileRegisterJSONRequest": {
        "properties": {
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/JudgeProfileInput"
            },
            "minItems": 1,
            "title": "Profiles",
            "type": "array"
          }
        },
        "required": [
          "profiles"
        ],
        "title": "JudgeProfileRegisterJSONRequest",
        "type": "object"
      },
      "JudgeProfileResponse": {
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "head_judge": {
            "title": "Head Judge",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "profile_prompt": {
            "title": "Profile Prompt",
            "type": "string"
          },
          "scoring_style": {
            "title": "Scoring Style",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "challenge_id",
          "domain",
          "scoring_style",
          "profile_prompt",
          "head_judge",
          "source_type",
          "created_at",
          "updated_at"
        ],
        "title": "JudgeProfileResponse",
        "type": "object"
      },
      "JudgeProfileURLRequest": {
        "properties": {
          "max_bytes": {
            "default": 1048576,
            "maximum": 5242880.0,
            "minimum": 1024.0,
            "title": "Max Bytes",
            "type": "integer"
          },
          "timeout_seconds": {
            "default": 10,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "type": "integer"
          },
          "url": {
            "minLength": 8,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "JudgeProfileURLRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Hackathon Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/challenges": {
      "post": {
        "operationId": "create_challenge_challenges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Challenge",
        "tags": [
          "challenges"
        ]
      }
    },
...
    "/challenges/{challenge_id}/judge-profiles/csv": {
      "post": {
        "operationId": "register_judge_profiles_csv_challenges__challenge_id__judge_profiles_csv_post",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/csv": {
              "schema": {
                "title": "Payload",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JudgeProfileResponse"
                  },
                  "title": "Response Register Judge Profiles Csv Challenges  Challenge Id  Judge Profiles Csv Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Judge Profiles Csv",
        "tags": [
          "judging"
        ]
      }
    },
    "/challenges/{challenge_id}/judge-profiles/json": {
      "post": {
        "operationId": "register_judge_profiles_json_challenges__challenge_id__judge_profiles_json_post",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JudgeProfileRegisterJSONRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JudgeProfileResponse"
                  },
                  "title": "Response Register Judge Profiles Json Challenges  Challenge Id  Judge Profiles Json Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Judge Profiles Json",
        "tags": [
          "judging"
        ]
      }
    },
    "/challenges/{challenge_id}/judge-profiles/url": {
      "post": {
        "operationId": "register_judge_profile_url_challenges__challenge_id__judge_profiles_url_post",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JudgeProfileURLRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JudgeProfileResponse"
                  },
                  "title": "Response Register Judge Profile Url Challenges  Challenge Id  Judge Profiles Url Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Judge Profile Url",
        "tags": [
          "judging"
        ]
      }
    },
    "/challenges/{challenge_id}/judge-profiles/yaml": {
      "post": {
        "operationId": "register_judge_profiles_yaml_challenges__challenge_id__judge_profiles_yaml_post",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-yaml": {
              "schema": {
                "title": "Payload",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JudgeProfileResponse"
                  },
                  "title": "Response Register Judge Profiles Yaml Challenges  Challenge Id  Judge Profiles Yaml Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Judge Profiles Yaml",
        "tags": [
          "judging"
        ]
      }
    },
...
  }
}